AIP-72: Move non-user facing code to `_internal`
Anything within `_internal` module is not a user-facing code, this makes it clearer! It won't be covered by Semantic versioning and can have breaking changes even in patch versions. A user should never use anything in `airflow/sdk/definitions/_internal`.
