+++ airflow/decorators/task_group.py
@@ -40,7 +40,7 @@ from airflow.models.expandinput import (
     MappedArgument,
 )
 from airflow.models.xcom_arg import XComArg
-from airflow.sdk.definitions.node import DAGNode
+from airflow.sdk.definitions._internal.node import DAGNode
 from airflow.typing_compat import ParamSpec
 from airflow.utils.helpers import prevent_duplicates
 from airflow.utils.task_group import MappedTaskGroup, TaskGroup
