+++ airflow/notifications/basenotifier.py
@@ -21,7 +21,7 @@ from abc import abstractmethod
 from collections.abc import Sequence
 from typing import TYPE_CHECKING
 
-from airflow.sdk.definitions.templater import Templater
+from airflow.sdk.definitions._internal.templater import Templater
 from airflow.utils.context import context_merge
 from airflow.utils.log.logging_mixin import LoggingMixin
