+++ task_sdk/src/airflow/sdk/definitions/_internal/__init__.py
@@ -14,15 +14,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-from __future__ import annotations
-
-from typing import TYPE_CHECKING
-
-if TYPE_CHECKING:
-    from airflow.typing_compat import TypeAlias
-
-import airflow.sdk.definitions._internal.mixins
-import airflow.sdk.definitions._internal.node
-
-DependencyMixin: TypeAlias = airflow.sdk.definitions._internal.mixins.DependencyMixin
-DAGNode: TypeAlias = airflow.sdk.definitions._internal.node.DAGNode
