+++ task_sdk/src/airflow/sdk/definitions/_internal/types.py
@@ -52,7 +52,7 @@ NOTSET = ArgNotSet()
 if TYPE_CHECKING:
     import logging
 
-    from airflow.sdk.definitions.node import DAGNode
+    from airflow.sdk.definitions._internal.node import DAGNode
 
     Logger = logging.Logger
 else:
